Falta implementar strategy 2 y reveer 1

+++ src/TrabajoPractico4_Queue/BankSim/Client.java
@@ -30,7 +30,7 @@ public class Client {
     public String getName() {
         return name;
     }
-    public double getEnterTime() {
+    public int getEnterTime() {
         return enterTime;
     }
     public void setEnterTime(int enterTime) {
